% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{addConvexHulls}
\alias{addConvexHulls}
\title{Overlay convex hulls onto a \code{plot2D} plot}
\usage{
addConvexHulls(object, fcol = "markers", ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}}

\item{fcol}{A \code{character} that defined which feature data
variable to be used as marker definition. Default is
\code{"markers"}.}

\item{...}{Additional parameters passed to `plot2D`.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
The function calculates and draws convex hulls on top of a plot
produced by \code{\link{plot2D}}. The hulls are computed in the
two componnets as defined by the \code{plot2D} method and
dimensions.
}
\examples{
library("pRolocdata")
data(E14TG2aS1)
plot2D(E14TG2aS1)
addConvexHulls(E14TG2aS1)
}
\author{
Laurent Gatto
}

