% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomarkers.R
\name{addGoMarkers}
\alias{addGoMarkers}
\title{Add GO markers}
\usage{
addGoMarkers(object, params, evidence, useID = FALSE, fcol = "GOMarkers",
  ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}.}

\item{params}{An instance of class \code{AnnotationParams}. 
If missing, \code{\link{getAnnotationParams}} will be used.}

\item{evidence}{GO evidence filtering.}

\item{useID}{Logical. Should GO term names or identifiers be used?
If \code{TRUE}, identifiers will be used. If \code{FALSE} GO term 
names will be used.}

\item{fcol}{Character. Name of the matrix of markers to be added to the 
\code{fData} default is \code{GOMarkers}}

\item{...}{Other arguments passed to \code{makeGoSet}}
}
\value{
An updated \code{MSnSet} with new feature data column
called \code{GOMarkers} containing a matrix of GO
markers
}
\description{
Adds a GO markers to the feature data
}
\examples{
library(pRolocdata)
data(dunkley2006)
setAnnotationParams(inputs =
                      c("Arabidopsis thaliana genes",
                                          "TAIR locus ID"))
xx <- addGoMarkers(dunkley2006)
dim(fData(xx)$GOMarkers)

## filter sets
xx <- filterMinMarkers(xx, n = 50)
dim(fData(xx)$GOMarkers)
xx <- filterMaxMarkers(xx, p = .25)
dim(fData(xx)$GOMarkers)

## Subset for specific protein sets
sub <- subsetMarkers(xx, keep = c("vacuole"))

## Order protein sets
res <- orderGoMarkers(xx, k = 1:3, p = 1/3, verbose = FALSE)
if (interactive()) {
pRolocVis(res, fcol = "GOMarkers")
}
}
\author{
Lisa M Breckels
}

