% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compfnames}
\alias{compfnames}
\title{compfnames}
\usage{
compfnames(obj1, obj2, fcol1 = "markers", fcol2, verbose = TRUE)
}
\arguments{
\item{obj1}{an object of class \code{"MSnSet"}.}

\item{obj2}{an object of class \code{"MSnSet"}.}

\item{fcol1}{an object of class \code{"character"} either in
\code{fvarLabels(obj1)} or \code{NULL}.}

\item{fcol2}{an object of class \code{"character"} either in
\code{fvarLabels(obj2)} or \code{NULL}.}

\item{verbose}{an object of class \code{"logical"}. Default \code{TRUE}.
Set to \code{FALSE} to suppress printing matrix.}
}
\value{
\code{compfnames} prints a matrix showing features common in
\code{obj1} and \code{obj2} and being unique to \code{obj1} and \code{obj2}.
This is done in total and for each unique marker level. Invisibly it
returns a list which contains feature names split accordingly to
\code{"all"} and to each unique marker level giving information about
features in common and unique for \code{obj1} and \code{obj2}.
}
\description{
\code{compfnames} allows to compare two objects of class
\code{"MSnSet"} with regard to their occurrence of features in total and
separated to unique marker levels on the basis of their \code{featureNames}.
Calling \code{compfnames} prints a
matrix giving information at first glance about similarities and
differences. Invisibly a list containing the individual \code{featureNames}
is returned.
}
\details{
\code{featureNames(obj1)} and \code{featureNames(obj2)} are used
to compute similarities and differences in the occurence of features.
\code{fcol1} and \code{fcol2} take measured variable names or \code{NULL} as
input which may be of great expedience in comparative spatial proteomics.
}
\author{
Laurent Gatto <lg390@cam.ac.uk>, Thomas Naake <tn299@cam.ac.uk>
}

