% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MSnSet.R
\name{filterBinMSnSet}
\alias{filterBinMSnSet}
\title{Filter a binary MSnSet}
\usage{
filterBinMSnSet(object, MARGIN = 2, t, q)
}
\arguments{
\item{object}{An \code{MSnSet}}

\item{MARGIN}{1 or 2. Default is 2.}

\item{t}{Absolute number of non-0 values to filer out columns (rows).}

\item{q}{Quantile defining the number of columns (rows) to filter.}
}
\value{
A filtered \code{MSnSet}.
}
\description{
Removes instances of columns or rows that have a certain
proportion or absolute number of 0 values.
}
\author{
Laurent Gatto
}

