% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MSnSet.R
\name{filterZeroCols}
\alias{filterZeroCols}
\title{Remove 0 columns}
\usage{
filterZeroCols(object, verbose = TRUE)
}
\arguments{
\item{object}{A \code{MSnSet} object.}

\item{verbose}{Print a message with the number of filtered out
columns (if any).}
}
\value{
An \code{MSnSet}.
}
\description{
Removes all assay data columns that are composed of only 0,
i.e. have a \code{colSum} of 0.
}
\examples{
library("pRolocdata")
data(andy2011goCC)
any(colSums(exprs(andy2011goCC)) == 0)
exprs(andy2011goCC)[, 1:5] <- 0
ncol(andy2011goCC)
ncol(filterZeroCols(andy2011goCC))
}
\author{
Laurent Gatto
}

