% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGoSet.R
\name{getGOFromFeatures}
\alias{getGOFromFeatures}
\title{Retrieve GO terms for feature names}
\usage{
getGOFromFeatures(id, namespace = "cellular_component", evidence = NULL,
  params = NULL, verbose = FALSE, nmax = 500)
}
\arguments{
\item{id}{An \code{character} with feature names to be pulled from
biomart. If and \code{MSnSet} is provided, then
\code{featureNames(id)} is used.}

\item{namespace}{The GO namespace. One of
\code{biological_process}, \code{cellular_component} (default)
or \code{molecular_function}.}

\item{evidence}{The GO evidence code. See
\code{showGOEvidenceCodes} for details. If \code{NULL}
(default), no filtering based on the evidence code is
performed.}

\item{params}{An instance of class
\code{"\linkS4class{AnnotationParams}"}.}

\item{verbose}{A \code{logical} defining verbosity of the
function. Default is \code{FALSE}.}

\item{nmax}{As described in
\url{https://support.bioconductor.org/p/86358/}, the Biomart
result can be unreliable for large queries. This argument
splits the input in chunks of length \code{nmax} (default is
500). If set to \code{NULL}, the query is performed in full.}
}
\value{
A \code{data.frame} with relevant GO terms.
}
\description{
The function pulls the gene ontology (GO) terms
for a set of feature names.
}
\examples{
library(pRolocdata)
data(dunkley2006)
data(dunkley2006params)
dunkley2006params
fn <- featureNames(dunkley2006)[1:5]
getGOFromFeatures(fn, params = dunkley2006params)
}
\author{
Laurent Gatto
}
