% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MSnSet.R
\name{getMarkerClasses}
\alias{getMarkerClasses}
\title{Returns the organelle classes in an 'MSnSet'}
\usage{
getMarkerClasses(object, fcol = "markers", verbose = TRUE, ...)
}
\arguments{
\item{object}{An instance of class \code{"\linkS4class{MSnSet}"}.}

\item{fcol}{The name of the markers column in the \code{featureData}
slot. Default is \code{markers}.}

\item{verbose}{If \code{TRUE}, a character vector of the organelle
classes is printed and the classes are returned invisibly. If \code{FALSE},
the markers are returned.}

\item{...}{Additional parameters passed to \code{sort} from the base package.}
}
\value{
A \code{character} vector of the organelle classes in the data.
}
\description{
Convenience accessor to the organelle classes in an 'MSnSet'.
This function returns the organelle classes of an
\code{MSnSet} instance. As a side effect, it prints out the classes.
}
\examples{
library("pRolocdata")
data(dunkley2006)
organelles <- getMarkerClasses(dunkley2006)
}
\author{
Lisa Breckels
}
\seealso{
\code{\link{getMarkers}}
}

