% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{geweke_test}
\alias{geweke_test}
\title{Wrapper for the geweke diagnostics from the coda package also
return p-values.}
\usage{
geweke_test(x)
}
\description{
Wrapper for the geweke diagnostics from the coda package also
return p-values.
}
