\name{knnPrediction}
\alias{knnPrediction}
\title{knn prediction}
\usage{
  knnPrediction(object, assessRes,
    scores = c("prediction", "all", "none"), k,
    fcol = "markers")
}
\arguments{
  \item{object}{An instance of class
  \code{"\linkS4class{MSnSet}"}.}

  \item{assessRes}{An instance of class
  \code{"\linkS4class{GenRegRes}"}, as generated by
  \code{\link{knnRegularisation}}.}

  \item{scores}{One of \code{"prediction"}, \code{"all"} or
  \code{"none"} to report the score for the predicted class
  only, for all cluster or none.}

  \item{k}{If \code{assessRes} is missing, a \code{k} must
  be provided.}

  \item{fcol}{The feature meta-data containing marker
  definitions. Default is \code{markers}.}
}
\value{
  An instance of class \code{"\linkS4class{MSnSet}"} with
  \code{knn} and \code{knn.scores} feature variables
  storing the classification results and scores
  respectively.
}
\description{
  Classification using for the k-nearest neighbours
  algorithm.
}
\examples{
library(pRolocdata)
data(dunkley2006)
## reducing parameter search space and interations
reg <- knnRegularisation(dunkley2006, k = c(3, 10), times = 3)
reg
plot(reg)
levelPlot(reg)
getRegularisedParams(reg)
res <- knnPrediction(dunkley2006, reg)
getPredictions(res, fcol = "knn")
getPredictions(res, fcol = "knn", t = 0.75)
plot2D(res, fcol = "knn")
}
\author{
  Laurent Gatto
}

