% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGoSet.R
\name{makeGoSet}
\alias{makeGoSet}
\title{Creates a GO feature \code{MSnSet}}
\usage{
makeGoSet(object, params, namespace = "cellular_component", evidence = NULL)
}
\arguments{
\item{object}{An instance of class \code{"\linkS4class{MSnSet}"}
or a character of feature names.}

\item{params}{An instance of class \code{"\linkS4class{AnnotationParams}"},
compatible with \code{featureNames(object)}'s format.}

\item{namespace}{The ontology name space. One or several of
\code{"biological_process"},  \code{"cellular_component"} or
\code{"molecular_function"}.}

\item{evidence}{GO evidence filtering.}
}
\value{
A new \code{"\linkS4class{MSnSet}"} with the GO terms
for the respective features in the original \code{object}.
}
\description{
Creates a new \code{"\linkS4class{MSnSet}"} instance populated
with a GO term binary matrix based on an original \code{object}.
}
\examples{
library("pRolocdata")
data(dunkley2006)
data(dunkley2006params)
goset <- makeGoSet(dunkley2006[1:10, ],
                   dunkley2006params)
goset
exprs(goset)[1:10, 1:5]
image(goset)
}
\author{
Laurent Gatto
}
