% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-framework.R
\docType{class}
\name{MAPparams-class}
\alias{MAPparams-class}
\alias{class:MAPparams}
\title{Class \code{"MAPparams"}}
\description{
TAGM parameters container.
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{Object of class \code{"character"} storing
the machine learning algorithm name.}

\item{\code{priors}}{Object of class \code{"list"} with the priors for the parameters}

\item{\code{design}}{Object of class \code{"numeric"} describing any cross-validation
parameters}

\item{\code{log}}{Object of class \code{"list"} with warnings}

\item{\code{seed}}{Object of class \code{"integer"} with the random
number generation seed.}

\item{\code{posteriors}}{Object of class \code{"list"} with the updated posterior parameters
and log-posterior of the model.}

\item{\code{datasize}}{Object of class \code{"list"} with details about size of data}
}}

