\name{MartInstance-class}
\Rdversion{1.1}
\docType{class}
\alias{MartInstance-class}
\alias{MartInstance}
\alias{show,MartInstance-method}

\alias{MartInstanceList-class}
\alias{MartInstanceList}

\alias{as.data.frame.MartInstanceList}
\alias{as.data.frame.MartInstance}

\alias{[,MartInstanceList-method}
\alias{[,MartInstanceList,ANY,ANY-method}
\alias{[,MartInstanceList,ANY,ANY,ANY-method}
\alias{[[,MartInstanceList-method}
\alias{[[,MartInstanceList,ANY,ANY-method}

\alias{sapply,MartInstanceList-method}
\alias{sapply,MartInstanceList,ANY-method}
\alias{lapply,MartInstanceList-method}
\alias{lapply,MartInstanceList,ANY-method}

\alias{nDatasets}
\alias{filterAttrs}
\alias{getMartInstanceList}
\alias{getMartTab}
\alias{getFilterList}

\title{Class \code{"MartInstance"}}

\description{
  
  Internal infrastructure to query/handle several individual mart
  instance. See \code{MartInterface.R} for details.     
}

\author{
  Laurent Gatto <lg390@cam.ac.uk>
}

