% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\docType{methods}
\name{plot,MCMCParams,character-method}
\alias{plot,MCMCParams,character-method}
\title{Plot posterior probabilities}
\usage{
\S4method{plot}{MCMCParams,character}(x, y, ...)
}
\arguments{
\item{x}{An object of class `MCMCParams`.}

\item{y}{A `character(1)` with a protein name.}

\item{...}{Currently ignored.}
}
\value{
A ggplot2 object.
}
\description{
Produces a violin plot with the protein posterior probabilities
distributions for all organelles.
}
