% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{mcmc_burn_chains}
\alias{mcmc_burn_chains}
\title{MCMC chain burning}
\usage{
mcmc_burn_chains(x, n = 50)
}
\arguments{
\item{x}{A object of class \code{MCMCParams}}

\item{n}{\code{integer(1)} defining number of iterations to burn. The default is
\code{50}}
}
\value{
An updated \code{MCMCParams} object.
}
\description{
Helper function to burn n iterations from the front of the chains
}
