% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{mcmc_get_meanComponent}
\alias{mcmc_get_meanComponent}
\title{Mean component allocation at each MCMC iteration}
\usage{
mcmc_get_meanComponent(x)
}
\arguments{
\item{x}{Object of class `MCMCParams`}
}
\value{
A `list` of length `length(x)`.
}
\description{
Helper function to get mean component allocation at each MCMC
iteration.
}
