% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{mcmc_get_meanoutliersProb}
\alias{mcmc_get_meanoutliersProb}
\title{Mean outlier probability}
\usage{
mcmc_get_meanoutliersProb(x)
}
\arguments{
\item{x}{Object of class `MCMCParams`}
}
\value{
A `list` of length `length(x)`.
}
\description{
Helper function to get mean probability of belonging to outlier at
each iteration.
}
\author{
Laurent Gatto
}
