% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{mcmc_get_outliers}
\alias{mcmc_get_outliers}
\title{Number of outlier at each iteration of MCMC}
\usage{
mcmc_get_outliers(x)
}
\arguments{
\item{x}{Object of class \code{MCMCParams}}
}
\value{
A \code{list} of length \code{length(x)}.
}
\description{
Helper function to get the number of outlier at each MCMC
iteration.
}
