% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{mcmc_pool_chains}
\alias{mcmc_pool_chains}
\title{Pool MCMC chains}
\usage{
mcmc_pool_chains(param)
}
\arguments{
\item{param}{An object of class \code{MCMCParams}.}
}
\value{
A pooled \code{MCMCParams} object.
}
\description{
Helper function to pool chains together after processing
}
