\name{nbPrediction}
\alias{nbPrediction}
\title{nb prediction}
\usage{
  nbPrediction(object, assessRes,
    scores = c("prediction", "all", "none"), laplace,
    fcol = "markers")
}
\arguments{
  \item{object}{An instance of class
  \code{"\linkS4class{MSnSet}"}.}

  \item{assessRes}{An instance of class
  \code{"\linkS4class{GenRegRes}"}, as generated by
  \code{\link{nbRegularisation}}.}

  \item{scores}{One of \code{"prediction"}, \code{"all"} or
  \code{"none"} to report the score for the predicted class
  only, for all cluster or none.}

  \item{laplace}{If \code{assessRes} is missing, a
  \code{laplace} must be provided.}

  \item{fcol}{The feature meta-data containing marker
  definitions. Default is \code{markers}.}
}
\value{
  An instance of class \code{"\linkS4class{MSnSet}"} with
  \code{nb} and \code{nb.scores} feature variables storing
  the classification results and scores respectively.
}
\description{
  Classification using the naive Bayes algorithm.
}
\examples{
library(pRolocdata)
data(dunkley2006)
## reducing parameter search space and iterations
reg <- nbRegularisation(dunkley2006, laplace = c(0, 5),  times = 3)
reg
plot(reg)
levelPlot(reg)
getRegularisedParams(reg)
res <- nbPrediction(dunkley2006, reg)
getPredictions(res, fcol = "naiveBayes")
getPredictions(res, fcol = "naiveBayes", t = 1)
plot2D(res, fcol = "naiveBayes")
}
\author{
  Laurent Gatto
}

