\name{nnetPrediction}
\alias{nnetPrediction}
\title{nnet prediction}
\usage{
  nnetPrediction(object, assessRes,
    scores = c("prediction", "all", "none"), decay, size,
    fcol = "markers")
}
\arguments{
  \item{object}{An instance of class
  code{"\linkS4class{MSnSet}"}.}

  \item{assessRes}{An instance of class
  code{"\linkS4class{GenRegRes}"}, as generated by
  \code{\link{nnetRegularisation}}.}

  \item{scores}{One of \code{"prediction"}, \code{"all"} or
  \code{"none"} to report the score for the predicted class
  only, for all cluster or none.}

  \item{decay}{If \code{assessRes} is missing, a
  \code{decay} must be provided.}

  \item{size}{If \code{assessRes} is missing, a \code{size}
  must be provided.}

  \item{fcol}{The feature meta-data containing marker
  definitions. Default is \code{markers}.}
}
\value{
  An instance of class code{"\linkS4class{MSnSet}"} with
  \code{nnet} and \code{nnet.scores} feature variables
  storing the classification results and scores
  respectively.
}
\description{
  Classification using the artificial neural network
  algorithm.
}
\author{
  Laurent Gatto
}

