% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-framework-tagm-novelty.R
\docType{class}
\name{MaxPear-class}
\alias{MaxPear-class}
\alias{.MaxPear}
\alias{NoveltyChain-class}
\alias{.NoveltyChain}
\alias{NoveltyChains-class}
\alias{.NoveltyChains}
\title{Contained for results of maximising the posterior expected adjusted
    rand index This class wraps the `list` ouput of `mcclust::maxpear` into
    an object.}
\description{
Contained for results of maximising the posterior expected adjusted
    rand index This class wraps the `list` ouput of `mcclust::maxpear` into
    an object.

Container for a single Novelty TAGM mcmc chain results

Contained for full Novelty TAGM results.
}
\section{Slots}{

\describe{
\item{\code{cl}}{A factor containing the organelle/phenotype to which the protein
has been assigned.}

\item{\code{value}}{numeric(1) containing the value of PEAR.}

\item{\code{method}}{character(1) describing the approach used to maximise PEAR. See
`mcclust::maxpear` for details.}

\item{\code{psm}}{A matrix where each entry is the posterior probability that
protein i localises with protein j}

\item{\code{maxPear}}{An instance of class \code{MaxPear} containing results from
maximising the Posterior Expected Adjusted Rand index.}

\item{\code{pooledNoveltyChain}}{An instance of class \code{NoveltyChain} after pooling
all separate chains.}

\item{\code{noveltyChains}}{\code{list()} containing the individual full Novelty TAGM
chain results in an \code{NoveltyChains} instance. Each element must be a
valid \code{NoveltyChain} instance.}

\item{\code{tagmNewclusterProb}}{A numeric vector containing the probability that
a protein belongs to a new cluster (aka phenotype).}
}}

