\name{plsdaPrediction}
\alias{plsdaPrediction}
\title{plsda prediction}
\usage{
  plsdaPrediction(object, assessRes,
    scores = c("prediction", "all", "none"), ncomp,
    fcol = "markers")
}
\arguments{
  \item{object}{An instance of class
  code{"\linkS4class{MSnSet}"}.}

  \item{assessRes}{An instance of class
  code{"\linkS4class{GenRegRes}"}, as generated by
  \code{\link{plsdaRegularisation}}.}

  \item{scores}{One of \code{"prediction"}, \code{"all"} or
  \code{"none"} to report the score for the predicted class
  only, for all cluster or none.}

  \item{ncomp}{If \code{assessRes} is missing, a
  \code{ncomp} must be provided.}

  \item{fcol}{The feature meta-data containing marker
  definitions. Default is \code{markers}.}
}
\value{
  An instance of class code{"\linkS4class{MSnSet}"} with
  \code{plsda} and \code{plsda.scores} feature variables
  storing the classification results and scores
  respectively.
}
\description{
  Classification using the partial least square
  distcriminant analysis algorithm.
}
\author{
  Laurent Gatto
}

