\name{rfPrediction}
\alias{rfPrediction}
\title{rf prediction}
\usage{
  rfPrediction(object, assessRes,
    scores = c("prediction", "all", "none"), mtry,
    fcol = "markers")
}
\arguments{
  \item{object}{An instance of class
  \code{"\linkS4class{MSnSet}"}.}

  \item{assessRes}{An instance of class
  \code{"\linkS4class{GenRegRes}"}, as generated by
  \code{\link{rfRegularisation}}.}

  \item{scores}{One of \code{"prediction"}, \code{"all"} or
  \code{"none"} to report the score for the predicted class
  only, for all cluster or none.}

  \item{mtry}{If \code{assessRes} is missing, a \code{mtry}
  must be provided.}

  \item{fcol}{The feature meta-data containing marker
  definitions. Default is \code{markers}.}
}
\value{
  An instance of class \code{"\linkS4class{MSnSet}"} with
  \code{rf} and \code{rf.scores} feature variables storing
  the classification results and scores respectively.
}
\description{
  Classification using the random forest algorithm.
}
\examples{
library(pRolocdata)
data(dunkley2006)
## reducing parameter search space and iterations
reg <- rfRegularisation(dunkley2006, mtry = c(2, 5, 10),  times = 3)
reg
plot(reg)
levelPlot(reg)
getRegularisedParams(reg)
res <- rfPrediction(dunkley2006, reg)
getPredictions(res, fcol = "rf")
getPredictions(res, fcol = "rf", t = 0.75)
plot2D(res, fcol = "rf")
}
\author{
  Laurent Gatto
}

