% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotation.R
\name{showGOEvidenceCodes}
\alias{getGOEvidenceCodes}
\alias{showGOEvidenceCodes}
\title{GO Evidence Codes}
\usage{
showGOEvidenceCodes()

getGOEvidenceCodes()
}
\value{
These functions are used for their side effects of printing
evidence codes and their description.
}
\description{
This function prints a textual description
of the Gene Ontology evidence codes.
}
\examples{
showGOEvidenceCodes()
getGOEvidenceCodes()
}
\author{
Laurent Gatto
}

