% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gomarkers.R
\name{subsetMarkers}
\alias{subsetMarkers}
\title{Subsets markers}
\usage{
subsetMarkers(object, fcol = "GOMarkers", keep)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}.}

\item{fcol}{The name of the markers matrix. Default is
\code{GOMarkers}.}

\item{keep}{Integer or character vector specifying the columns to keep 
in the markers matrix, as defined by \code{fcol}.}
}
\value{
An updated \code{MSnSet}
}
\description{
Subsets a matrix of markers by specific terms
}
\author{
Lisa M Breckels
}
\seealso{
\code{addGoMarkers} and example therein.
}

