\name{svmPrediction}
\alias{svmPrediction}
\title{ksvm prediction}
\usage{
  svmPrediction(object, assessRes,
    scores = c("prediction", "all", "none"), cost, sigma,
    fcol = "markers")
}
\arguments{
  \item{object}{An instance of class
  \code{"\linkS4class{MSnSet}"}.}

  \item{assessRes}{An instance of class
  \code{"\linkS4class{GenRegRes}"}, as generated by
  \code{\link{svmRegularisation}}.}

  \item{scores}{One of \code{"prediction"}, \code{"all"} or
  \code{"none"} to report the score for the predicted class
  only, for all cluster or none.}

  \item{cost}{If \code{assessRes} is missing, a \code{cost}
  must be provided.}

  \item{sigma}{If \code{assessRes} is missing, a
  \code{sigma} must be provided.}

  \item{fcol}{The feature meta-data containing marker
  definitions. Default is \code{markers}.}
}
\value{
  An instance of class \code{"\linkS4class{MSnSet}"} with
  \code{svm} and \code{svm.scores} feature variables
  storing the classification results and scores
  respectively.
}
\description{
  Classification using the support vector machine
  algorithm.
}
\examples{
library(pRolocdata)
data(dunkley2006)
## reducing parameter search space and interations
reg <- svmRegularisation(dunkley2006, cost = 2^seq(-2,2,2), sigma = 10^seq(-1, 1, 1),  times = 3)
reg
plot(reg)
levelPlot(reg)
getRegularisedParams(reg)
res <- svmPrediction(dunkley2006, reg)
getPredictions(res, fcol = "svm")
getPredictions(res, fcol = "svm", t = 0.75)
plot2D(res, fcol = "svm")
}
\author{
  Laurent Gatto
}

