% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-map.R
\name{tagmPredict}
\alias{tagmPredict}
\title{TAGM model for Mass spectrometry spatial proteomics datasets prediction routine using 
empirical Bayes methods (MAP).}
\usage{
tagmPredict(object, MAPparams, fcol = "markers",
  probreturn = c("prediction", "joint"), proboutlier = TRUE)
}
\arguments{
\item{object}{An instance of class \code{"\linkS4class{MSnSet}"}}

\item{MAPparams}{An instance of class \code{"\linkS4class{MAPparams}"}, as generated by
\code{\link{tagmTrain}}.}

\item{fcol}{The feature meta-data containing marker definitions.
Default is \code{markers}}

\item{probreturn}{One of \code{"prediction"}, \code{"joint"} to report the probability 
for the predicted class only, for all classes,}

\item{proboutlier}{Boolean indicating whether to return the probability of being an outlier}
}
\value{
An instance of class \code{"\linkS4class{MSnSet}"}
}
\description{
TAGM model for Mass spectrometry spatial proteomics datasets prediction routine using 
empirical Bayes methods (MAP).
}
