% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-map.R
\name{tagmTrain}
\alias{tagmTrain}
\title{TAGM MAP parameter estimation}
\usage{
tagmTrain(object, fcol = "markers", method = "MAP", numIter = 100,
  mu0 = NULL, lambda0 = 0.01, nu0 = NULL, S0 = NULL, beta0 = NULL,
  u = 2, v = 10, seed = NULL)
}
\arguments{
\item{object}{An instance of class \code{"\linkS4class{MSnSet}"}}

\item{fcol}{The feature meta-data containing marker definitions.
Default is \code{markers}}

\item{method}{The current infernece method for the TAGM algorithm. Default is \code{"MAP"}}

\item{numIter}{The number of iterations of the expectation-maximisation algorithm. Default is \code{100}}

\item{mu0}{The prior mean. Default is colmeans of expression data.}

\item{lambda0}{The prior shrinkage. Default is 0.01}

\item{nu0}{The prior degreed of freedom. Default is \code{ncol(exprs(object))+2}}

\item{S0}{The prior inverse-wishary scale matrix. Empirical prior used by default.}

\item{beta0}{The prior Dirichlet distribution concentration. Default is 1 for each class}

\item{u}{The prior shape parameter for Beta(u, v). Default is 2}

\item{v}{The prior shape parameter for Beta(u, v). Default is 10.}

\item{seed}{The optional random number generator seed.}
}
\value{
An instance of class \code{"\linkS4class{MAPparams}"}.
}
\description{
T augmented Gaussian mixture (TAGM) model for Mass spectrometry spatial proteomics datasets Optimisation routine using 
empirical Bayes methods (MAP).
}
\author{
Oliver M. Crook
}
