% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/machinelearning-functions-theta.R
\name{thetaClassification}
\alias{thetaClassification}
\title{theta classification}
\usage{
thetaClassification(primary, auxiliary, fcol = "markers", bestTheta, k,
  scores = c("prediction", "all", "none"))
}
\arguments{
\item{primary}{An instance of class \code{"\linkS4class{MSnSet}"}.}

\item{auxiliary}{An instance of class \code{"\linkS4class{MSnSet}"}.}

\item{fcol}{The feature meta-data containing marker definitions.
Default is \code{markers}.}

\item{bestTheta}{Best theta vector as output from
\code{thetaOptimisation}, see \code{thetaOptimisation} for details}

\item{k}{Numeric vector of length 2, containing the best \code{k}
parameters to use for the primary and auxiliary datasets. If k \code{k}
is not specified it will be calculated internally.}

\item{scores}{One of \code{"prediction"}, \code{"all"} or
\code{"none"} to report the score for the predicted class only,
for all cluster or none.}
}
\value{
A character vector of the classifications for the unknowns
}
\description{
Classification using a variation of the KNN implementation
of Wu and Dietterich's transfer learning schema
}
\author{
Lisa Breckels
}
\seealso{
\code{\link{thetaOptimisation}} and example therein.
}

