% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/machinelearning-functions-theta.R
\name{thetaOptimisation}
\alias{thetaOptimisation}
\alias{thetaOptimization}
\title{theta parameter optimisation}
\usage{
thetaOptimisation(primary, auxiliary, fcol = "markers", k, times = 20,
  test.size = 0.2, xval = 5, by = 0.5, length.out, th, xfolds,
  BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{primary}{An instance of class \code{"\linkS4class{MSnSet}"}.}

\item{auxiliary}{An instance of class \code{"\linkS4class{MSnSet}"}.}

\item{fcol}{The feature meta-data containing marker definitions.
Default is \code{markers}.}

\item{k}{Numeric vector of length 2, containing the best \code{k}
parameters to use for the primary (\code{k[1]}) and auxiliary
(\code{k[2]}) datasets. See \code{knnOptimisation} for generating
best \code{k}.}

\item{times}{The number of times cross-validation is
performed. Default is 20.}

\item{test.size}{The size of test (validation) data. Default is
0.2 (20 percent).}

\item{xval}{The number of rounds of cross-validation to perform.}

\item{by}{The increment for theta, must be one of \code{c(1, 0.5,
0.25, 0.2, 0.15, 0.1, 0.05)}}

\item{length.out}{Alternative to using \code{by}
parameter. Specifies the desired length of the sequence of theta
to test.}

\item{th}{A matrix of theta values to test for each class as
generated from the function \code{\link{thetas}}, the number of
columns should be equal to the number of classes contained in
\code{fcol}. Note: columns will be ordered according to
\code{getMarkerClasses(primary, fcol)}.}

\item{xfolds}{Option to pass specific folds for the cross
validation.}

\item{BPPARAM}{Required for parallelisation. If not specified
selects a default \code{BiocParallelParam}, from global options
or, if that fails, the most recently registered() back-end.}
}
\value{
A list of containing the theta combinations tested,
associated macro F1 score and accuracy for each combination over
each round (specified by times).
}
\description{
Classification parameter optimisation for the KNN implementation
of Wu and Dietterich's transfer learning schema
}
\details{
\code{thetaOptimisation} implements a variation of Wu and
Dietterich's transfer learning schema: P. Wu and
T. G. Dietterich. Improving SVM accuracy by training on auxiliary
data sources. In Proceedings of the Twenty-First International
Conference on Machine Learning, pages 871 - 878.  Morgan Kaufmann,
2004. A grid search for the best theta is performed.
}
\author{
Lisa Breckels
}
\seealso{
\code{\link{thetaClassification}} and example therein.
}

