\name{undocumented}
\alias{undocumented}

\alias{getParams,ClustRegRes-method}
\alias{levelPlot,ClustRegRes-method}
\alias{plot,ClustRegRes,missing-method}
\alias{show,ClustRegRes-method}

\title{Undocumented/unexported entries}

\description{
  This is just a dummy entry for methods from unexported classes that
  generate warnings during package checking. 
}


\author{
  Laurent Gatto <lg390@cam.ac.uk>
}
