% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{addLegend}
\alias{addLegend}
\title{Adds a legend}
\usage{
addLegend(object, fcol = "markers", where = c("bottomleft", "bottom",
  "bottomright", "left", "topleft", "top", "topright", "right", "center",
  "other"), col, bty = "n", ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}}

\item{fcol}{Feature meta-data label (fData column name) defining
the groups to be differentiated using different
colours. Default is \code{markers}.}

\item{where}{One of \code{"bottomleft"} (default),
\code{"bottomright"}, \code{"topleft"}, \code{"topright"} or
\code{"other"} defining the location of the
legend. \code{"other"} opens a new graphics device, while the
other locations are passed to \code{\link{legend}}.}

\item{col}{A \code{character} defining point colours.}

\item{bty}{Box type, as in \code{legend}. Default is set to
\code{"n"}.}

\item{...}{Additional parameters passed to \code{\link{legend}}.}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Adds a legend to a \code{\link{plot2D}} figure.
}
\details{
The function has been updated in version 1.3.6 to recycle the
default colours when more organelle classes are provided. See
\code{\link{plot2D}} for details.
}
\author{
Laurent Gatto
}
