% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterdist-functions.R
\name{getNormDist}
\alias{getNormDist}
\title{Extract Distances from a \code{"ClustDistList"} object}
\usage{
getNormDist(object, p = 1/3)
}
\arguments{
\item{object}{An instance of class \code{"\linkS4class{ClustDistList}"}.}

\item{p}{The normalisation factor. Default is 1/3.}
}
\value{
An numeric of normalised distances, one per protein set in the
\code{ClustDistList}.
}
\description{
This function computes and outputs normalised distances from a
\code{"\linkS4class{ClustDistList}"} object.
}
\seealso{
\code{"\linkS4class{ClustDistList}"}, \code{"\linkS4class{ClustDist}"},
and examples in \code{clustDist}.
}
\author{
Lisa Breckels
}
