\name{getStockcol}
\alias{getStockcol}
\alias{getStockpch}
\alias{getUnknowncol}
\alias{getUnknownpch}
\title{Default cluster colours}
\usage{
  getStockcol()

  getStockpch()

  getUnknowncol()

  getUnknownpch()
}
\value{
  A \code{character} vector.

  A \code{numeric} vector.

  A \code{character} vector or length 1.

  A \code{numeric} vector of length 1.
}
\description{
  Returns the default cluster colours.

  Returns the default cluster point symbols.

  Returns the default colour for unknown features.

  Returns the default point character for unknown features.
}
\examples{
getStockcol()
getStockpch()
getUnknowncol()
getUnknownpch()
}
\author{
  Laurent Gatto
}

