% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{setLisacol}
\alias{getLisacol}
\alias{getStockcol}
\alias{getStockpch}
\alias{getUnknowncol}
\alias{getUnknownpch}
\alias{setLisacol}
\alias{setStockcol}
\alias{setStockcolGui}
\alias{setStockpch}
\alias{setUnknowncol}
\alias{setUnknownpch}
\title{Manage default colours and point characters}
\usage{
setLisacol()

getLisacol()

getStockcol()

setStockcol(cols)

getStockpch()

setStockpch(pchs)

getUnknowncol()

setUnknowncol(col)

getUnknownpch()

setUnknownpch(pch)

setStockcolGui()
}
\arguments{
\item{cols}{A vector of colour \code{characters} or \code{NULL},
which sets the colours to the default values.}

\item{pchs}{A vector of \code{numeric} or \code{NULL},
which sets the point characters to the default values.}

\item{col}{A colour \code{character} or \code{NULL},
which sets the colour to \code{#E7E7E7} (\code{grey91}),
the default colour for unknown features.}

\item{pch}{A \code{numeric} vector of length 1 or \code{NULL},
which sets the point character to 21, the default.}
}
\value{
The \code{set} functions set (and invisibly returns)
colours. The \code{get} functions returns a \code{character}
vector of colours. For the \code{pch} functions, \code{numeric}s
rather than \code{character}s.
}
\description{
These functions allow to get/set the default colours and point
character that are used when plotting organelle clusters and
unknown features. These values are parametrised at the session
level. Two palettes are available: the default and Lisa's colours.
}
\examples{
## defaults for clusters
getStockcol()
getStockpch()
## unknown features
getUnknownpch()
getUnknowncol()
## an example
library(pRolocdata)
data(dunkley2006)
par(mfrow = c(2, 1))
plot2D(dunkley2006, fcol = "markers", main = 'Default colours')
setUnknowncol("black")
plot2D(dunkley2006, fcol = "markers", main = 'setUnknowncol("black")')
getUnknowncol()
setUnknowncol(NULL)
getUnknowncol()
getLisacol()
setLisacol()
## now default colours are Lisa's
getStockcol()
}
\author{
Laurent Gatto
}

