% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/go.R
\name{goIdToTerm}
\alias{flipGoTermId}
\alias{goIdToTerm}
\alias{goTermToId}
\alias{prettyGoTermId}
\title{Convert GO ids to/from terms}
\usage{
goIdToTerm(x)

goTermToId(x)

flipGoTermId(x)

prettyGoTermId(x)
}
\arguments{
\item{x}{A \code{character} of GO ids or terms.}
}
\value{
A \code{character} of GO terms (ids) if \code{x} were ids
(terms).
}
\description{
Converts GO identifiers to/from GO terms, either explicitly or by
checking if (any items in) the input contains \code{"GO:"}.
}
\examples{
goIdToTerm("GO:0000001")
goIdToTerm(c("GO:0000001", "novalid"))
goIdToTerm(c("GO:0000001", "GO:0000002", "notvalid"))
goTermToId("mitochondrion inheritance")
goTermToId(c("mitochondrion inheritance", "notvalid"))
prettyGoTermId("mitochondrion inheritance")
prettyGoTermId("GO:0000001")
flipGoTermId("mitochondrion inheritance")
flipGoTermId("GO:0000001")
}
\author{
Laurent Gatto
}

