% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go.R
\name{goIdToTerm}
\alias{goIdToTerm}
\alias{goTermToId}
\alias{flipGoTermId}
\alias{prettyGoTermId}
\title{Convert GO ids to/from terms}
\usage{
goIdToTerm(x, names = TRUE, keepNA = TRUE)

goTermToId(x, names = TRUE, keepNA = TRUE)

flipGoTermId(x, names = TRUE, keepNA = TRUE)

prettyGoTermId(x)
}
\arguments{
\item{x}{A \code{character} of GO ids or terms.}

\item{names}{Should a named character be returned? Default is
\code{TRUE}.}

\item{keepNA}{Should any GO term/id names that are missing or obsolete
be replaced with a \code{NA}? Default is \code{TRUE}. If 
\code{FALSE} then the GO term/id names is kept.}
}
\value{
A \code{character} of GO terms (ids) if \code{x} were ids
(terms).
}
\description{
Converts GO identifiers to/from GO terms, either explicitly or by
checking if (any items in) the input contains \code{"GO:"}.
}
\examples{
goIdToTerm("GO:0000001")
goIdToTerm("GO:0000001", names = FALSE)
goIdToTerm(c("GO:0000001", "novalid"))
goIdToTerm(c("GO:0000001", "GO:0000002", "notvalid"))
goTermToId("mitochondrion inheritance")
goTermToId("mitochondrion inheritance", name = FALSE)
goTermToId(c("mitochondrion inheritance", "notvalid"))
prettyGoTermId("mitochondrion inheritance")
prettyGoTermId("GO:0000001")
flipGoTermId("mitochondrion inheritance")
flipGoTermId("GO:0000001")
flipGoTermId("GO:0000001", names = FALSE)
}
\author{
Laurent Gatto
}
