% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{mcmc_burn_chains}
\alias{mcmc_burn_chains}
\title{MCMC chain burning}
\usage{
mcmc_burn_chains(x, n)
}
\arguments{
\item{x}{A object of class `MCMCParams`}

\item{n}{`integer(1)` defining number of iterations to burn.}
}
\value{
An updated `MCMCParams` object.
}
\description{
Helper function to burn n iterations from the front of the chains
}
\author{
Laurent Gatto
}
