% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-mcmc-helper.R
\name{mcmc_thin_chains}
\alias{mcmc_thin_chains}
\title{MCMC chain thinning}
\usage{
mcmc_thin_chains(x, freq)
}
\arguments{
\item{x}{An object of class `MCMCParams`.}

\item{freq}{Thinning frequency.}
}
\description{
Helper function to subsample the chains, known informally as
thinning.
}
\author{
Laurent Gatto
}
