\name{plsdaPrediction}
\alias{plsdaPrediction}
\title{plsda prediction}
\usage{
  plsdaPrediction(object, assessRes,
    scores = c("prediction", "all", "none"), ncomp,
    fcol = "markers")
}
\arguments{
  \item{object}{An instance of class
  \code{"\linkS4class{MSnSet}"}.}

  \item{assessRes}{An instance of class
  \code{"\linkS4class{GenRegRes}"}, as generated by
  \code{\link{plsdaRegularisation}}.}

  \item{scores}{One of \code{"prediction"}, \code{"all"} or
  \code{"none"} to report the score for the predicted class
  only, for all cluster or none.}

  \item{ncomp}{If \code{assessRes} is missing, a
  \code{ncomp} must be provided.}

  \item{fcol}{The feature meta-data containing marker
  definitions. Default is \code{markers}.}
}
\value{
  An instance of class \code{"\linkS4class{MSnSet}"} with
  \code{plsda} and \code{plsda.scores} feature variables
  storing the classification results and scores
  respectively.
}
\description{
  Classification using the partial least square
  distcriminant analysis algorithm.
}
\examples{
library(pRolocdata)
data(dunkley2006)
## reducing parameter search space and interations
reg <- plsdaRegularisation(dunkley2006, ncomp = c(3, 10),  times = 2)
reg
plot(reg)
levelPlot(reg)
getRegularisedParams(reg)
res <- plsdaPrediction(dunkley2006, reg)
getPredictions(res, fcol = "plsda")
getPredictions(res, fcol = "plsda", t = 0.75)
plot2D(res, fcol = "plsda")
}
\author{
  Laurent Gatto
}

