% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machinelearning-functions-tagm-map.R
\name{tagmTrain}
\alias{tagmTrain}
\alias{tagmPredict}
\title{TAGM MAP parameter estimation}
\usage{
tagmTrain(object, fcol = "markers", method = "MAP", numIter = 100,
  mu0 = NULL, lambda0 = 0.01, nu0 = NULL, S0 = NULL, beta0 = NULL,
  u = 2, v = 10, seed = NULL)

tagmPredict(object, params, fcol = "markers", probreturn = c("prediction",
  "joint"), proboutlier = TRUE)
}
\arguments{
\item{object}{An \code{\link[MSnbase:MSnSet]{MSnbase::MSnSet}} containing the spatial
proteomics data.}

\item{fcol}{The feature meta-data containing marker definitions.
Default is \code{markers}.}

\item{method}{A \code{charachter()} describing the inference method for
the TAGM algorithm. Default is \code{"MAP"}.}

\item{numIter}{The number of iterations of the
expectation-maximisation algorithm. Default is 100.}

\item{mu0}{The prior mean. Default is \code{colMeans} of the expression
data.}

\item{lambda0}{The prior shrinkage. Default is 0.01.}

\item{nu0}{The prior degreed of freedom. Default is
\code{ncol(exprs(object)) + 2}}

\item{S0}{The prior inverse-wishary scale matrix. Empirical prior
used by default.}

\item{beta0}{The prior Dirichlet distribution
concentration. Default is 1 for each class.}

\item{u}{The prior shape parameter for Beta(u, v). Default is 2}

\item{v}{The prior shape parameter for Beta(u, v). Default is 10.}

\item{seed}{The optional random number generator seed.}

\item{params}{An instance of class \code{\link{MAPparams}}, as generated by
\code{\link[=tagmTrain]{tagmTrain()}}.}

\item{probreturn}{One of \code{"prediction"} or \code{"joint"} to report the
probability for the predicted class only (as a new
\code{tagm.probability} feature variable), or for all classes (as a
new \code{tagm.joint} feature variable).}

\item{proboutlier}{A \code{logical(1)} indicating whether to return the
probability of being an outlier as a new \code{tagm.outlier}
feature variable.}
}
\value{
An instance of class \code{\link[=MAPparams]{MAPparams()}}.

An instance of class \code{\link[MSnbase:MSnSet]{MSnbase::MSnSet}} containing the
localisation predictions as a new \code{tagm.allocation} feature
variable.
}
\description{
T augmented Gaussian mixture (TAGM) model for mass
spectrometry-based spatial proteomics datasets optimisation
routine using empirical Bayes methods (MAP).
}
\references{
\emph{A Bayesian Mixture Modelling Approach For Spatial
Proteomics} Oliver M Crook, Claire M Mulvey, Paul D. W. Kirk,
Kathryn S Lilley, Laurent Gatto bioRxiv 282269; doi:
https://doi.org/10.1101/282269
}
\author{
Oliver M. Crook
}
